requirements.txt
gradio==5.9.1
ultralytics==8.3.58
tensorflow==2.17.1
opencv-python==4.10.0
Pillow==11.0.0
numpy==1.26.4
gdown==5.2.0
